Add files via upload
I made code that can merge both of the APIs. We would need a range of dates for each week of the stock data. By comparing the dates, we can add the attributes of the YahooAPI to the Perigon article data. 

I was able to use .explode() to duplicate rows that had more than one article featured. Now, every row only features a single company.